+++ packages/Rfmalloc/src/fmalloc.cpp
@@ -12,5 +12,6 @@
 #include "fmalloc_bed.inc"
 #include "fmalloc_dosage.inc"
 #include "fmalloc_hap.inc"
+#include "fmalloc_ld.inc"
 #include "fmalloc_insitu.inc"
 #include "fmalloc_api.inc"